requirements.txt
azure-functions
azure-storage-blob
smbprotocol
logging

